requirements.txt
pymupdf
sentence-transformers
faiss-cpu
langchain
langchain-groq
langchain-community
pydantic>=2
fastapi
uvicorn[standard]
python-multipart
python-dotenv
